% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai.R
\docType{data}
\name{ai}
\alias{ai}
\title{A dataset with artificial insemination data in 100 herds}
\format{
A data frame with 96 786 rows and 4 variables:
\describe{
  \item{herd_id}{herd ID}
  \item{cow_id}{cow ID}
  \item{ai_date}{date of artificial insemination}
  \item{bull_id}{ID of the bull used for the insemination}
}
}
\usage{
ai
}
\description{
A dataset with artificial insemination data in 100 herds
}
\keyword{datasets}
