% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anim.R
\docType{data}
\name{anim}
\alias{anim}
\title{A dataset with individual information on the animals in the different datasets.}
\format{
A data frame with 12 907 rows and 6 variables:
\describe{
  \item{anim_id}{Animal ID}
  \item{birth_date}{Date of birth}
  \item{birth_herd}{ID of the herd in which the animal was born}
  \item{dam_id}{ID of the animal dam}
  \item{bull_id}{ID of the animal sire}
  \item{breed}{Code for the animal breed}
}
}
\usage{
anim
}
\description{
A dataset with individual information on the animals in the different datasets.
}
\keyword{datasets}
